Share createTreeNodeInput function
