+++ ui/components-react/src/test/tree/controlled/component/TreeNodeEditor.test.tsx
@@ -8,7 +8,7 @@ import sinon from "sinon";
 import { act, fireEvent, render, waitFor } from "@testing-library/react";
 import { TreeNodeEditor } from "../../../../components-react/tree/controlled/component/TreeNodeEditor";
 import { MutableTreeModelNode } from "../../../../components-react/tree/controlled/TreeModel";
-import { createRandomMutableTreeModelNode } from "../RandomTreeNodesHelpers";
+import { createRandomMutableTreeModelNode } from "../TreeHelpers";
 
 describe("TreeNodeEditor", () => {
   let testNode: MutableTreeModelNode;
