+++ ui/components-react/src/test/tree/controlled/internal/SparseTree.test.ts
@@ -6,7 +6,7 @@ import { expect } from "chai";
 import * as faker from "faker";
 import sinon from "sinon";
 import { Node, SparseArray, SparseTree } from "../../../../components-react/tree/controlled/internal/SparseTree";
-import { createRandomMutableTreeModelNode, createRandomMutableTreeModelNodes } from "../RandomTreeNodesHelpers";
+import { createRandomMutableTreeModelNode, createRandomMutableTreeModelNodes } from "../TreeHelpers";
 
 describe("SparseTree", () => {
   interface TestNode extends Node {
